requirements.txt
websocket-client==1.2.1
ccxt==1.58.9
numpy==1.21.1
ta==0.7.0